requirements.txt
streamlit==1.49.1
tensorflow==2.20.0
pillow==11.3.0
numpy==2.2.6
huggingface_hub==0.34.4
